move encode_seal to root of the crate
